lstsize_return_int

+++ ft_lstsize.c
@@ -6,7 +6,7 @@
 /*   By: ktakada <ktakada@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/04/13 16:41:39 by ktakada           #+#    #+#             */
-/*   Updated: 2022/04/13 16:50:26 by ktakada          ###   ########.fr       */
+/*   Updated: 2022/04/13 23:24:18 by ktakada          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_lstsize(t_list *lst)
 {
-	size_t	count;
+	int	count;
 
 	count = 0;
 	while (lst != NULL)
